refactor(data): extract repository CRUD operations into separate module
