+++ src/data/src/db/infused.rs
@@ -1,4 +1,3 @@
-
 use std::collections::HashMap;
 use std::ops::Deref;
 use std::str::FromStr;
