+++ src/data/src/db/mod.rs
@@ -29,6 +29,9 @@ pub enum FoundOrInserted<T: Table> {
 /// Repository module
 pub mod repo;
 
+/// Repository module (mostly just CRUD)
+pub mod repo_crud;
+
 /// Entities with extra information embedded.
 pub mod infused;
